Create AttributeComponent type in preparation for refactoring
