+++ client/anoixo-client/src/query/Attributes/AttributeComponent.ts
@@ -1,7 +1,7 @@
 /**
  * Props type for components for editing word query attributes.
  */
-type AttributeComponentProps = {
+export type AttributeComponentProps = {
   /** The key for the attribute in the word query's `attributes` dictionary */
   id: string;
   /** The current value for the attribute in the word query's `attributes` dictionary, if there is one */
@@ -12,4 +12,4 @@ type AttributeComponentProps = {
   enabled: boolean;
 };
 
-export default AttributeComponentProps;
+export type AttributeComponent = React.FC<AttributeComponentProps>;
